feat(DonationInfo Constructor): Abstracting DonationInfo  to set members dynamically
Making DonationInfo constructor take only a donation in Donation, abstracting it in case a donation is missing a field.

Changing DonationInfo to be more abstract - taking a donation and dynamically setting the fields, iwhile making sure that if a field is missing - adding it with an empty string.

Chaning field names in DonationsFilter to match the fields given in the request.