feat(donation creation in the database): Making connection to the database and creating a Donation node
Renaming DonationInfo to DonationDomain.

Creating DonationsGraph class to make the connection to the neo4j database.

Creating in [main.py](http://main.py) a create_donation method to create a donation node in the database.

Creating a models directory with DonationModel class, containing a mapping between a DonationDomain and the Donation node in the db, and the create function.

Creating a repositories directory with DonationRepository class, containing a create method that calls the DonationModel.create method.